requirements.txt
# Django Framework specific
Django==1.9.5
django-environ==0.4.0
djangorestframework==3.3.3
drfdocs==0.0.9
gunicorn==19.4.5

# Serve static files
whitenoise==3.0

psycopg2==2.6.1
six==1.10.0
ipython==4.1.2

# Testing
flake8==2.5.4
coverage==4.0.3
django-autofixture==0.12.1
factory_boy==2.7.0
pygithub==1.25.2
